Full Functioning Live Serial Plotter
The "Grapher.cs" script was updated in which is now able able input an "n" number of points and plot the line connections between each point, live. When the number of points is updated and/or the Duty Cycle is set, the graph adjusts accordingly and in-game(live). The next step is to create a slide/toggle to change the duty cycle in-game to show live serial plotting of the Arduino.